tests: Fix REQUESTED_* handling and remove redundant logic
Fix incorrect use of `IN LISTS` when iterating over hardcoded variable
suffixes introduced in edaa1e3 ("tests: Add script-mode tests with support
for requested and expected variables", 2025-07-26). Use `IN ITEMS` instead
to correctly process REQUESTED_* inputs.

With this fix in place, the earlier individual `if()` blocks become
redundant and are removed.